Fixed bug in the (new) Wn and Zn functions.
Return variable f is declared static. Idea was to prevent this piece of memory from being allocated and deallocated unnecessarily, the thousands of times that it is called during execution. This may be unnecessary but current thinking is it does not hurt.

Pitfall: When doing this, variable must be initialized in a separate statement (f=0.0;), otherwise initialization only happens on the first function call and every subsequent evaluation is wrong. Bug was: (static Double_t f=0.0;) when should have used (static Double_t f;) and (f=0.0;).

 Please enter the commit message for your changes. Lines starting