New definition of Y2. Takes 8 minutes just to draw all functions! Also the output is not correct.
I have been using 'for' loops to handle complicated sums, rather than finding the closed forms. This appears not to be tractable. The loop for Y2 calls the loops for W, Z, and X pops. This will only get worse for Y3. Committing this as a waypoint. Like I said, Y2 is still not even correct, much less evaluating quickly.