Y2 much closer. Y3 coded. 120sec to draw.
Have caught a large number of bugs. Highlight: typing trapping efficiency parameter as 'p' instead of 'a[p]'. Results in evaluting 5 instead of ~0.75. That one took a while.

Y3 is now coded. Both Y2 and Y3 are lacking the feeding terms in background interval, so their shapes are not right. The trapping solutions are still off somehow, but they have gotten much closer.

Big issue ahead: I believe the fitter is doing a large number of complex calculations redundantly. Any quantity that depends on the parameters but not the t[0] variable should be calculated only once per step in parameter space. But all my pre-computations of these quantities are happening inside the functions. I would like to insert a function for precomputing these values and storing them in global variables, to be accessed during evaluation. These quantities include:

1) Initial values. This is just a single evaluation of the trapping-interval solution.
2) The Sigma's -- sums over a number of teeth. These can be put into an array with, say, 200 elements.
3) Modified lifetimes
4) Convert a[gamma]'s to 1/ms
5) Complex coefficients

As part of this, evaluation of the trapping solution should have its own function. This will get called once to find the initial condition and then many times for evalations at other points. (Right now the initial condition is being called for every point. This should gian a factor of two in performance.