New Y2InitialValue() & Y3InitialValue() functions.
The difficult part of the work should live in these functions. The full solution should be constructable from these in a straightforward way.

Previously the background solution was not properly constructed, and this makes that easy.

I can't remember what made the Y2 population fit better.... :/ but since what I'm doing now is simpler I'm confident it is actually better.