Works on BMC_0005! Does not work on BMC_0009.
BMC_0005 is the more complete and complex case. It is puzzling that it doesn't fit BMC_0009.

BFit has been updated to include a variable number of points (nPoints) for evaluating functions. Previously it was always 1 point per millisecond, eg. 100,000's in the case of 137-I.

BFitModel is very hacky right now. Several versions of Y3InitialValue are commented out. There are choices made about how to handle the background and initial values of parent populations that I can't fully justify. There is more work to do.

Execution time is still a major problem. Doing the regression on BMC_0005 took 1182 seconds.