Refactored how Y2 and Y3 are calculated.
Previously called and summed a Y2/Y3InitialValue function for each tooth. Now I do the integral up to tooth n-1 for each n and store these in an array. These arrays are now accessed in Ycap and Y2/Y3InitialValue is no longer needed. I did this because I had a problem with the W contributions to Y2 and Y3 and didn't know how to solve it.

The work is not complete. See the Ycap function: So far have V, W, Z, and X contributions to Y3 working. Right now there's a problem with V, W, and Z to Y2. Then I need to do Y to Y3.